+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## FIWARE Big Bang v0.37.0-next
 
+-   Update IoT Agent for UltraLight to 3.4.0 (#370)
+-   Update STH Comet to 2.11.0 (#369)
 -   Improve maintenance script (#368)
 -   Update Node-RED to 3.1.9 (#367)
 -   Support Ubuntu 24.04 (#366)